+++ Source/GameLoop/Game.cpp
@@ -39,10 +39,14 @@ void Game::Start() {
 	ui->CrateTempFontFromTTF("Fonts/arial.ttf", 8, "arial8");
 
 	renderer->FLatRenderCopySetUp();
+
+	cam = new Camera(false);
 }
 
 void Game::LogicUpdate() {
 	Global::frameCounter++;
+	const Uint8* state = SDL_GetKeyboardState(nullptr);
+	cam->UpdatePosition(state);
 }
 
 void Game::FrameUpdate() {
@@ -54,6 +58,7 @@ void Game::FrameUpdate() {
 void Game::Input() {
 	while (SDL_PollEvent(&event)) {
 		ui->ManageInput(event);
+		cam->UpdateZoom(event);
 		Exit();
 	}
 	Global::inputDelay++;
@@ -63,8 +68,16 @@ void Game::Render() {
 	renderer->ClearFrame(255, 255, 255);
 	MT::Rect rect{ 200,200,100,100 };
 	Point mousePos = GetMousePos();
-	renderer->RenderShadowUPR(rect, TexMan::GetTex("tree1"), {0,0,0,120},{(float)mousePos.x,(float)mousePos.y,0.5f});
-	renderer->RenderCopyUPR(rect, TexMan::GetTex("tree1"));
+	MT::Rect rect1{ 10,10,100,100 };
+	MT::Rect rect2{ 300,300,100,100 };
+	MT::Rect rect3{ 500,500,20,20 };
+	MT::Texture* tex = TexMan::GetTex("tree1");
+	rect1 = cam->Transform(rect1);
+	rect2 = cam->Transform(rect2);
+	rect3 = cam->Transform(rect3);
+	renderer->RenderCopy(rect1, tex);
+	renderer->RenderCopy(rect2, tex);
+	renderer->RenderCopy(rect3, tex);
 	ui->Render();
 	renderer->Present();
 }